+++ ICD.Connect.Routing.CrestronPro/Cards/Outputs/DmcHdo/DmcHdoAdapter.cs
@@ -1,6 +1,9 @@
+using System;
 using System.Collections.Generic;
 using System.Linq;
+using ICD.Connect.Misc.CrestronPro.Devices;
 #if SIMPLSHARP
+using Crestron.SimplSharpPro;
 using Crestron.SimplSharpPro.DM;
 using Crestron.SimplSharpPro.DM.Cards;
 #endif
@@ -9,7 +12,7 @@ namespace ICD.Connect.Routing.CrestronPro.Cards.Outputs.DmcHdo
 {
 #if SIMPLSHARP
 	// ReSharper disable once InconsistentNaming
-	public sealed class DmcHdoAdapter : AbstractOutputCardAdapter<DmcHdoSingle, DmcHdoAdapterSettings>
+	public sealed class DmcHdoAdapter : AbstractOutputCardAdapter<DmcHdoSingle, DmcHdoAdapterSettings>, IPortParent
 	{
 		/// <summary>
 		/// Constructor.
@@ -51,7 +54,91 @@ namespace ICD.Connect.Routing.CrestronPro.Cards.Outputs.DmcHdo
 		{
 			return new DmcHdoSingle(cardNumber, switcher);
 		}
+
+		#region IPortParent
+
+		/// <summary>
+		/// Gets the port at the given address.
+		/// </summary>
+		/// <param name="address"></param>
+		/// <returns></returns>
+		public ComPort GetComPort(int address)
+		{
+			string message = string.Format("{0} has no {1}", this, typeof(ComPort).Name);
+			throw new NotSupportedException(message);
+		}
+
+		/// <summary>
+		/// Gets the port at the given address.
+		/// </summary>
+		/// <param name="address"></param>
+		/// <returns></returns>
+		public IROutputPort GetIrOutputPort(int address)
+		{
+			string message = string.Format("{0} has no {1}", this, typeof(IROutputPort).Name);
+			throw new NotSupportedException(message);
+		}
+
+		/// <summary>
+		/// Gets the port at the given address.
+		/// </summary>
+		/// <param name="address"></param>
+		/// <returns></returns>
+		public Relay GetRelayPort(int address)
+		{
+			string message = string.Format("{0} has no {1}", this, typeof(Relay).Name);
+			throw new NotSupportedException(message);
+		}
+
+		/// <summary>
+		/// Gets the port at the given address.
+		/// </summary>
+		/// <param name="address"></param>
+		/// <returns></returns>
+		public Versiport GetIoPort(int address)
+		{
+			string message = string.Format("{0} has no {1}", this, typeof(Versiport).Name);
+			throw new NotSupportedException(message);
+		}
+
+		/// <summary>
+		/// Gets the port at the given address.
+		/// </summary>
+		/// <param name="address"></param>
+		/// <returns></returns>
+		public DigitalInput GetDigitalInputPort(int address)
+		{
+			string message = string.Format("{0} has no {1}", this, typeof(DigitalInput).Name);
+			throw new NotSupportedException(message);
+		}
+
+		/// <summary>
+		/// Gets the port at the given address.
+		/// </summary>
+		/// <param name="io"></param>
+		/// <param name="address"></param>
+		/// <returns></returns>
+		public Cec GetCecPort(eInputOuptut io, int address)
+		{
+			if (io == eInputOuptut.Output)
+			{
+				switch (address)
+				{
+					case 1:
+						return Card.Card1.HdmiOutput.StreamCec;
+					case 2:
+						return Card.Card2.HdmiOutput.StreamCec;
+				}
+			}
+
+			string message = string.Format("{0} has no {1} at address {2}:{3}", this, typeof(Cec).Name, io, address);
+			throw new ArgumentException(message);
+		}
+
+		#endregion
 	}
+
+	
 #else
 	public sealed class DmcHdoAdapter : AbstractOutputCardAdapter<DmcHdoAdapterSettings>
 	{