+++ ICD.Connect.Routing.Atlona/Properties/AssemblyInfo.cs
@@ -4,4 +4,4 @@ using System.Reflection;
 [assembly: AssemblyCompany("ICD Systems")]
 [assembly: AssemblyProduct("ICD.Connect.Routing.Atlona")]
 [assembly: AssemblyCopyright("Copyright © ICD Systems 2018")]
-[assembly: AssemblyVersion("14.0.0.0")]
+[assembly: AssemblyVersion("14.1.0.0")]