+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [14.1.0] - 2019-01-29
 ### Added
  - DMC-SO-HD DM 8G Fiber Output Card
  - DM-RMC-200-S DM 8G Fiber Receiver